% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odstostd.R
\name{main}
\alias{main}
\title{Title ods工资/社保合并工时到std}
\usage{
main(token)
}
\arguments{
\item{token}{口令}
}
\value{
合并后结果
}
\description{
Title ods工资/社保合并工时到std
}
\examples{
main(token='057A7F0E-F187-4975-8873-AF71666429AB')

}
